Remove Admin-only and generic presenter related code from Opportunity presenter
