+++ spec/presenters/base_presenter_spec.rb
@@ -9,6 +9,26 @@ RSpec.describe BasePresenter do
     end
   end
 
+  describe '#put' do
+    it 'returns value when present' do
+      presenter = BasePresenter.new
+
+      expect(presenter.put('some present text')).to eq('some present text')
+    end
+
+    it 'returns default when value not present' do
+      presenter = BasePresenter.new
+
+      expect(presenter.put('')).to eq('none')
+    end
+
+    it 'returns specified default when value not present' do
+      presenter = BasePresenter.new
+
+      expect(presenter.put('', 'something else')).to eq('something else')
+    end
+  end
+
   # Private methods
 
   describe '::h' do
